@@ -1,49 +1,43 @@
 source 'https://rubygems.org'
-gem 'rails', '3.2.16'
 
-gem 'sqlite3'
-
+# Backend
+gem 'rails', '4.1.4'
+gem 'pg'
 gem 'devise'
-gem 'cancan'
-
-gem 'jquery-rails'
-gem 'simple_form'
-
+gem 'cancancan'
 gem 'carrierwave'
 gem 'mini_magick'
-
-gem 'ransack'
-gem 'squeel'
-gem 'kaminari'
-gem 'rocket_tag'
-
-gem 'factory_girl_rails'
-
-gem 'capistrano'
-gem 'rvm-capistrano'
+gem 'ransack', github: 'activerecord-hackery/ransack', branch: 'rails-4.1'
 gem 'whenever', require: false
-
-gem 'flickraw'
-gem 'disqus'
 gem 'nokogiri'
 
-gem 'coveralls', require: false
+# Frontend
+gem 'jquery-rails'
+gem 'simple_form'
+gem 'kaminari'
 
 group :assets do
-  gem 'sass-rails'
+  gem 'sass-rails', '4.0.3'
   gem 'coffee-rails'
   gem 'uglifier'
   gem 'therubyracer'
 end
 
+group :development do
+  gem 'capistrano'
+  gem 'rvm-capistrano'
+  gem 'quiet_assets'
+end
+
 group :development, :test do
-  gem 'debugger'
-  gem 'thin'
+  gem 'byebug'
 end
 
 group :test do
+  gem 'factory_girl_rails'
+  gem 'coveralls', require: false
   gem 'rspec-rails'
-  gem 'database_cleaner'
+  gem 'database_rewinder'
   gem 'shoulda'
 end
 
