@@ -8,7 +8,7 @@ describe Ability do
       @ability = Ability.new(@user)
     end
     it 'should have user abilities' do
-      @ability.should be_able_to(:manage, :all)
+      expect(@ability.can?(:manage, :all)).to eq true
     end
   end
 
@@ -18,23 +18,23 @@ describe Ability do
     end
 
     it 'cannot manage all' do
-      @ability.should_not be_able_to(:manage, :all)
+      expect(@ability).to_not be_able_to(:manage, :all)
     end
 
     it 'can :read published Album' do
-      @ability.should be_able_to(:read, Album.new(is_published: true))
+      expect(@ability).to be_able_to(:read, Album.new(is_published: true))
     end
 
     it 'cannot :read non-published Album' do
-      @ability.should_not be_able_to(:read, Album.new(is_published: false))
+      expect(@ability).to_not be_able_to(:read, Album.new(is_published: false))
     end
 
     it 'can :read published Project' do
-      @ability.should be_able_to(:read, Project.new(is_published: true))
+      expect(@ability).to be_able_to(:read, Project.new(is_published: true))
     end
 
     it 'cannot :read non-published Project' do
-      @ability.should_not be_able_to(:read, Project.new(is_published: false))
+      expect(@ability).to_not be_able_to(:read, Project.new(is_published: false))
     end
   end
 end
