@@ -1,10 +1,6 @@
 require 'spec_helper'
 
-describe AlbumsController do
-  it 'should use AlbumsController' do
-    controller.should be_an_instance_of(AlbumsController)
-  end
-
+describe AlbumsController, type: :controller do
   describe 'unauthorized request' do
     context 'accessible pages' do
       after :each do
@@ -13,16 +9,16 @@ describe AlbumsController do
       it 'should be success' do
         @album = FactoryGirl.create(:album, images: [FactoryGirl.create(:image)])
         get :show, id: @album.id
-        response.status.should eq(200)
-        response.should render_template(:show)
+        expect(response.status).to eq(200)
+        expect(response).to render_template(:show)
       end
     end
 
     context 'inaccessible pages' do
       context 'albums' do
         after :each do
-          response.should redirect_to root_path
-          response.status.should eq(302)
+          expect(response).to redirect_to root_path
+          expect(response.status).to eq(302)
         end
 
         it 'should redirect to homepage' do
@@ -32,18 +28,18 @@ describe AlbumsController do
           get :new
         end
         it 'should redirect to homepage' do
-          post :create
+          post :create, { album: { title: '' } }
         end
       end
 
       context 'members' do
-        before :each do
+        before do
           @album = FactoryGirl.create(:album)
         end
 
         after :each do
-          response.should redirect_to root_path
-          response.status.should eq(302)
+          expect(response).to redirect_to root_path
+          expect(response.status).to eq(302)
         end
 
         it 'should redirect to homepage' do
@@ -60,70 +56,70 @@ describe AlbumsController do
   end
 
   context 'authorized request' do
-    before :each do
+    before do
       @user = FactoryGirl.create(:user)
       sign_in @user
     end
 
     describe "GET 'new'" do
-      before :each do
+      before do
         get :new
       end
 
       it 'should be successful' do
-        response.should be_success
-        response.should render_template(:new)
+        expect(response).to be_success
+        expect(response).to render_template(:new)
       end
     end
 
     describe "POST 'create'" do
-      before :each do
+      before do
         post :create, album: { title: 'aa AA Aa aA', description: 'Bb' }
       end
 
       it 'should be successful' do
-        response.status.should eq(302)
-        Album.last.present?.should be_true
-        Album.last.title.should eq('aa AA Aa aA')
-        Album.last.description.should eq('Bb')
+        expect(response.status).to eq 302
+        album = Album.last
+        expect(album.title).to eq 'aa AA Aa aA'
+        expect(album.description).to eq 'Bb'
       end
     end
 
     describe "GET 'edit'" do
-      before :each do
+      before do
         @album = FactoryGirl.create(:album)
         get :edit, id: @album.id
       end
 
       it 'should be successful' do
-        response.should be_success
-        response.should render_template(:edit)
+        expect(response.status).to eq 200
+        expect(response).to render_template(:edit)
       end
     end
 
     describe "PUT 'update'" do
-      before :each do
+      before do
         @album = FactoryGirl.create(:album)
         put :update, id: @album.id, album: { title: 'BB bb Bb bB!', description: 'Bb' }
       end
 
       it 'should be successful' do
-        response.status.should eq(302)
+        expect(response.status).to eq 302
         @album.reload
-        @album.title.should eq('BB bb Bb bB!')
-        @album.description.should eq('Bb')
+        expect(@album.title).to eq('BB bb Bb bB!')
+        expect(@album.description).to eq('Bb')
       end
     end
 
     describe "DELETE 'destroy'" do
-      before :each do
+      before do
         @album = FactoryGirl.create(:album)
         delete :destroy, id: @album.id
       end
 
       it 'should be successful' do
-        response.status.should eq(302)
-        Album.find_by_id(@album.id).nil?.should be_true
+        expect(response.status).to eq 302
+        expect(Album.find_by_id(@album.id)).to eq nil
       end
     end
   end
