@@ -1,25 +1,21 @@
 require 'spec_helper'
 
-describe HomeController do
-  it 'should use HomeController' do
-    controller.should be_an_instance_of(HomeController)
-  end
-
+describe HomeController, type: :controller do
   describe "GET 'index'" do
     it 'should be successful' do
       get :index
-      response.should be_success
-      response.should render_template(:index)
-      response.content_type.should eq('text/html')
+      expect(response).to be_success
+      expect(response).to render_template(:index)
+      expect(response.content_type).to eq('text/html')
     end
   end
 
   describe "GET 'rss'" do
     it 'should be successful' do
       get :rss, format: :rss
-      response.should be_success
-      response.should render_template('home/rss')
-      response.content_type.should eq('application/rss+xml')
+      expect(response).to be_success
+      expect(response).to render_template('home/rss')
+      expect(response.content_type).to eq('application/rss+xml')
     end
   end
 end
