@@ -14,27 +14,27 @@ describe ImageHelper do
     describe 'check_title' do
       it 'should render correct title suggestion' do
         @image.is_for_sale = false
-        check_title(@image).to_s.should match("#{Image::TITLE_MIN} words")
+        expect(check_title(@image).to_s).to match("#{Image::TITLE_MIN} words")
         @image.is_for_sale = true
-        check_title(@image).to_s.should match("#{Image::TITLE_MIN_FOR_SALE} words")
+        expect(check_title(@image).to_s).to match("#{Image::TITLE_MIN_FOR_SALE} words")
       end
     end
 
     describe 'check_desc' do
       it 'should render correct desc suggestion' do
         @image.is_for_sale = false
-        check_desc(@image).to_s.should match("#{Image::DESC_MIN} words")
+        expect(check_desc(@image).to_s).to match("#{Image::DESC_MIN} words")
         @image.is_for_sale = true
-        check_desc(@image).to_s.should match("#{Image::DESC_MIN_FOR_SALE} words")
+        expect(check_desc(@image).to_s).to match("#{Image::DESC_MIN_FOR_SALE} words")
       end
     end
 
     describe 'check_tags' do
       it 'should render correct desc suggestion' do
         @image.is_for_sale = false
-        check_tags(@image).to_s.should match("#{Image::TAGS_MIN} tags")
+        expect(check_tags(@image).to_s).to match("#{Image::TAGS_MIN} tags")
         @image.is_for_sale = true
-        check_tags(@image).to_s.should match("#{Image::TAGS_MIN_FOR_SALE} tags")
+        expect(check_tags(@image).to_s).to match("#{Image::TAGS_MIN_FOR_SALE} tags")
       end
     end
   end
