@@ -3,9 +3,8 @@ class ImagesController < ApplicationController
   load_resource :image, only: [:show]
 
   def index
-    @q = Image.unscoped.includes([:album]).search(params[:q])
+    @q = Image.sorted.includes([:album]).search(params[:q])
     @images = @q.result
-    @images = @images.order('id DESC') if params[:q].nil?
     @images = @images.page(params[:page])
   end
 
@@ -27,7 +26,7 @@ class ImagesController < ApplicationController
   end
 
   def update
-    if @image.update_attributes(params[:image])
+    if @image.update_attributes(image_params)
       redirect_to image_path(@image)
     else
       render :edit
@@ -40,5 +39,15 @@ class ImagesController < ApplicationController
   def destroy
     @image.destroy
     redirect_to images_path
+  end
+
+private
+
+  def image_params
+    params.require(:image).permit(
+      :asset, :asset_cache, :album_id, :title, :title_ua, :desc, :desc_ua, :place, :place_ua, :date,
+      :updated_at, :published_at_checkbox, :tags, :tags_resolved,
+      :is_for_sale, :image_width, :image_height
+    )
   end
 end
