@@ -1,9 +1,9 @@
 require 'spec_helper'
 
-describe PhotosController do
+describe PhotosController, type: :routing do
   describe 'routing' do
     it 'recognizes and generates CRUD' do
-      pending
+      expect(delete: '/photos/1').to route_to(controller: 'photos', action: 'destroy', id: '1')
     end
   end
 end
